filter_service/requirements.txt
aio_pika==9.5.5
aiohttp==3.11.16
boto3==1.37.37
botocore==1.37.37
numpy==2.2.6
opencv_python==4.11.0.86
Pillow==11.2.1
python-dotenv==1.1.0
redis==5.2.1
tenacity==9.1.2
